flight

+++ plane-service/src/cpp/models/FlightModel.cpp
@@ -1,6 +1,6 @@
 #include "../../header/models/FlightModel.h"
 
-FlightModel::FlightModel(int id, long int timestampstart, long int timestampend, int dispatcherId, int planeId, int airportId)
+FlightModel::FlightModel(long int id, long int timestampstart, long int timestampend, long int dispatcherId, long int planeId, long int airportId)
 {
     _id = id;
     _timstampstart = timestampstart;
@@ -10,7 +10,7 @@ FlightModel::FlightModel(int id, long int timestampstart, long int timestampend,
     _airportId = airportId;
 }
 
-int FlightModel::getId()
+long int FlightModel::getId()
 {
     return _id;
 }
@@ -22,20 +22,20 @@ long int FlightModel::getTimestampEnd()
 {
     return _timestampend;
 }
-int FlightModel::getDispatcherId()
+long int FlightModel::getDispatcherId()
 {
     return _dispatcherId;
 }
-int FlightModel::getPlaneId()
+long int FlightModel::getPlaneId()
 {
     return _planeId;
 }
-int FlightModel::getAirportId()
+long int FlightModel::getAirportId()
 {
     return _airportId;
 }
 
-void FlightModel::setId(int id)
+void FlightModel::setId(long int id)
 {
     _id = id;
 }
@@ -47,15 +47,15 @@ void FlightModel::setTimestampEnd(long int timestampend)
 {
     _timestampend = timestampend;
 }
-void FlightModel::setDispatcherId(int dispatcherId)
+void FlightModel::setDispatcherId(long int dispatcherId)
 {
     _dispatcherId = dispatcherId;
 }
-void FlightModel::setPlaneId(int planeId)
+void FlightModel::setPlaneId(long int planeId)
 {
     _planeId = planeId;
 }
-void FlightModel::setAirportId(int airportId)
+void FlightModel::setAirportId(long int airportId)
 {
     _airportId = airportId;
 }
